+++ pageserver/src/walreceiver.rs
@@ -150,6 +150,7 @@ fn walreceiver_main(
         error!("No previous WAL position");
     }
 
+    // FIXME: We have to do it to handle new segment generated by pg_resetwal at compute node startup
     startpoint = Lsn::max(
         startpoint,
         Lsn(end_of_wal.0 & !(pg_constants::WAL_SEGMENT_SIZE as u64 - 1)),
